--- categories_dict.py
@@ -1,1397 +1,0 @@
-CATEGORIES = {
-  "Rutracker Awards (мероприятия и конкурсы)": [
-    "Конкурсы",
-    "Rutracker Awards (Раздачи)"
-  ],
-  "Зарубежное кино": [
-    "Классика мирового кинематографа",
-    "Фильмы до 1990 года",
-    "Фильмы 1991-2000",
-    "Фильмы 2001-2005",
-    "Фильмы 2006-2010",
-    "Фильмы 2011-2015",
-    "Фильмы 2016-2018",
-    "Фильмы 2019",
-    "Фильмы Ближнего Зарубежья",
-    "Азиатские фильмы",
-    "Индийское кино",
-    "Сборники фильмов",
-    "Короткий метр",
-    "Грайндхаус",
-    "Звуковые дорожки и Переводы"
-  ],
-  "Наше кино": [
-    "Кино СССР",
-    "Детские отечественные фильмы",
-    "Авторские дебюты"
-  ],
-  "Арт-хаус и авторское кино": [
-    "Короткий метр (Арт-хаус и авторское кино)",
-    "Документальные фильмы (Арт-хаус и авторское кино)",
-    "Анимация (Арт-хаус и авторское кино)"
-  ],
-  "Театр": [],
-  "DVD Video": [
-    "Классика мирового кинематографа (DVD Video)",
-    "Азиатские фильмы (DVD Video)",
-    "Зарубежное кино (DVD Video)",
-    "Наше кино (DVD Video)",
-    "Фильмы Ближнего Зарубежья (DVD Video)",
-    "Арт-хаус и авторское кино (DVD Video)",
-    "Индийское кино (DVD Video)",
-    "Грайндхаус (DVD Video)"
-  ],
-  "HD Video": [
-    "UHD Video",
-    "Классика мирового кинематографа (HD Video)",
-    "Зарубежное кино (HD Video)",
-    "Азиатские фильмы (HD Video)",
-    "Наше кино (HD Video)",
-    "Арт-хаус и авторское кино (HD Video)",
-    "Индийское кино (HD Video)",
-    "Грайндхаус (HD Video)"
-  ],
-  "3D/Стерео Кино, Видео, TV и Спорт": [
-    "3D Кинофильмы",
-    "3D Мультфильмы",
-    "3D Документальные фильмы",
-    "3D Спорт",
-    "3D Ролики, Музыкальное видео, Трейлеры к фильмам"
-  ],
-  "Мультфильмы": [
-    "Отечественные мультфильмы (HD Video)",
-    "Иностранные мультфильмы (HD Video)",
-    "Иностранные короткометражные мультфильмы (HD Video)",
-    "Отечественные мультфильмы (DVD)",
-    "Иностранные мультфильмы (DVD)",
-    "Иностранные короткометражные мультфильмы (DVD)",
-    "Отечественные мультфильмы",
-    "Отечественные полнометражные мультфильмы",
-    "Иностранные мультфильмы",
-    "Иностранные короткометражные мультфильмы",
-    "Сборники мультфильмов"
-  ],
-  "Мультсериалы": [
-    "Мультсериалы (SD Video)",
-    "Мультсериалы (DVD Video)",
-    "Мультсериалы (HD Video)"
-  ],
-  "Аниме": [
-    "Артбуки и журналы (Аниме)",
-    "Обои, сканы, аватары, арт",
-    "AMV и другие ролики",
-    "Аниме (DVD)",
-    "Аниме (HD Video)",
-    "Аниме (основной подраздел)",
-    "Аниме (плеерный подраздел)",
-    "Аниме (QC подраздел)",
-    "Покемоны",
-    "Наруто",
-    "Гандам",
-    "Японские мультфильмы",
-    "Звуковые дорожки (Аниме)"
-  ],
-  "Русские сериалы": [
-    "Возвращение Мухтара",
-    "Воронины",
-    "Глухарь / Пятницкий / Карпов",
-    "Земский доктор",
-    "Каменская",
-    "Кухня / Отель Элеон",
-    "Ментовские войны",
-    "Молодежка / Интерны",
-    "Морские дьяволы",
-    "Москва. Три вокзала",
-    "Нюхач",
-    "Обратная сторона Луны",
-    "Ольга / Физрук",
-    "Пуля Дура",
-    "Сваты",
-    "След",
-    "Солдаты и пр.",
-    "Тайны следствия",
-    "Улицы разбитых фонарей (Менты) / Опера / Убойная сила"
-  ],
-  "Зарубежные сериалы": [
-    "Новинки и сериалы в стадии показа",
-    "Сериалы США и Канады",
-    "Сериалы Великобритании и Ирландии",
-    "Скандинавские сериалы",
-    "Испанские сериалы",
-    "Итальянские сериалы",
-    "Европейские сериалы",
-    "Сериалы стран Африки, Ближнего и Среднего Востока",
-    "Сериалы Австралии и Новой Зеландии",
-    "Сериалы Ближнего Зарубежья",
-    "Сериалы совместного производства нескольких стран",
-    "Веб-сериалы, Вебизоды к сериалам и Пилотные серии сериалов",
-    "Анатомия Грей + Частная Практика",
-    "Бесстыжие / Shameless (US)",
-    "Вавилон 5 / Babylon 5",
-    "Викинги / Vikings",
-    "Во все тяжкие / Breaking Bad",
-    "Дневники вампира + Настоящая кровь",
-    "Доктор Кто + Торчвуд",
-    "Доктор Хаус / House M.D.",
-    "Друзья + Джоуи",
-    "Звёздные Врата : Атлантида; Вселенная",
-    "Звёздные Врата: СГ1 / Stargate: SG1",
-    "Звёздный крейсер Галактика + Каприка",
-    "Звёздный путь / Star Trek",
-    "Игра престолов / Game of Thrones",
-    "Карточный Домик / House of Cards",
-    "Клан Сопрано / The Sopranos",
-    "Кости / Bones",
-    "Менталист + Касл",
-    "Место преступления / CSI: Crime Scene Investigation",
-    "Морская полиция: Спецотдел; Лос-Анджелес; Новый Орлеан",
-    "Оранжевый — хит сезона / Orange Is the New Black",
-    "Остаться в Живых / LOST",
-    "Отчаянные домохозяйки / Desperate Housewives",
-    "Побег из тюрьмы / Prison Break",
-    "Сверхъестественное / Supernatural",
-    "Секретные материалы / The X-Files",
-    "Секс в большом городе / Sex And The City",
-    "Твин пикс / Twin Peaks",
-    "Теория большого взрыва / The Big Bang Theory",
-    "Форс-мажоры / Костюмы в законе / Suits",
-    "Ходячие мертвецы + Бойтесь ходячих мертвецов",
-    "Черное зеркало / Black Mirror",
-    "Для некондиционных раздач"
-  ],
-  "Зарубежные сериалы (HD Video)": [
-    "Викинги / Vikings (HD Video)",
-    "Друзья / Friends (HD Video)",
-    "Доктор Кто + Торчвуд (HD Video)",
-    "Доктор Хаус / House M.D. (HD Video)",
-    "Звёздные Врата (HD Video)",
-    "Звёздный крейсер Галактика + Каприка (HD Video)",
-    "Звёздный путь / Star Trek (HD Video)",
-    "Игра престолов / Game of Thrones (HD Video)",
-    "Карточный домик (HD Video)",
-    "Кости / Bones (HD Video)",
-    "Менталист + Касл (HD Video)",
-    "Место преступления / CSI: Crime Scene Investigation (HD Video)",
-    "Оранжевый — хит сезона / Orange Is the New Black (HD Video)",
-    "Остаться в Живых / LOST (HD Video)",
-    "Побег из тюрьмы / Prison Break (HD Video)",
-    "Сверхъестественное / Supernatural (HD Video)",
-    "Секретные материалы / The X-Files (HD Video)",
-    "Твин пикс / Twin Peaks (HD Video)",
-    "Теория Большого Взрыва / The Big Bang Theory (HD Video)",
-    "Ходячие мертвецы + Бойтесь ходячих мертвецов (HD Video)",
-    "Черное зеркало / Black Mirror (HD Video)",
-    "Для некондиционных раздач (HD Video)"
-  ],
-  "Сериалы Латинской Америки, Турции и Индии": [
-    "Актёры и актрисы латиноамериканских сериалов",
-    "Сериалы Аргентины",
-    "Сериалы Бразилии",
-    "Сериалы Венесуэлы",
-    "Сериалы Индии",
-    "Сериалы Колумбии",
-    "Сериалы Латинской Америки с озвучкой (раздачи папками)",
-    "Сериалы Латинской Америки с субтитрами",
-    "Официальные краткие версии сериалов Латинской Америки",
-    "Сериалы Мексики",
-    "Сериалы Перу, Сальвадора, Чили и других стран",
-    "Сериалы совместного производства",
-    "Сериалы США (латиноамериканские)",
-    "Сериалы Турции",
-    "Для некондиционных раздач",
-    "OST Сериалы Латинской Америки, Турции и Индии (lossy и lossless)"
-  ],
-  "Азиатские сериалы": [
-    "Китайские сериалы с субтитрами",
-    "Корейские сериалы с озвучкой",
-    "Корейские сериалы с субтитрами",
-    "Прочие азиатские сериалы с озвучкой",
-    "Тайваньские сериалы с субтитрами",
-    "Японские сериалы с субтитрами",
-    "Японские сериалы с озвучкой",
-    "VMV и др. ролики",
-    "OST Азиатские сериалы (lossy и lossless)"
-  ],
-  "Вера и религия": [
-    "[Видео Религия] Христианство",
-    "[Видео Религия] Ислам",
-    "[Видео Религия] Религии Индии, Тибета и Восточной Азии",
-    "[Видео Религия] Культы и новые религиозные движения"
-  ],
-  "Документальные фильмы и телепередачи": [
-    "Документальные (DVD)",
-    "[Док] Биографии. Личности и кумиры",
-    "[Док] Кинематограф и мультипликация",
-    "[Док] Мастера искусств Театра и Кино",
-    "[Док] Искусство, история искусств",
-    "[Док] Музыка",
-    "[Док] Криминальная документалистика",
-    "[Док] Тайны века / Спецслужбы / Теории Заговоров",
-    "[Док] Военное дело",
-    "[Док] Вторая мировая война",
-    "[Док] Аварии / Катастрофы / Катаклизмы",
-    "[Док] Авиация",
-    "[Док] Космос",
-    "[Док] Научно-популярные фильмы",
-    "[Док] Флора и фауна",
-    "[Док] Путешествия и туризм",
-    "[Док] Медицина",
-    "[Док] Социальные ток-шоу",
-    "[Док] Информационно-аналитические и общественно-политические перед..",
-    "[Док] Архитектура и строительство",
-    "[Док] Всё о доме, быте и дизайне",
-    "[Док] BBC",
-    "[Док] Discovery",
-    "[Док] National Geographic",
-    "[Док] История: Древний мир / Античность / Средневековье",
-    "[Док] История: Новое и Новейшее время",
-    "[Док] Эпоха СССР",
-    "[Док] Битва экстрасенсов / Теория невероятности / Искатели / Галил..",
-    "[Док] Русские сенсации / Программа Максимум / Профессия репортёр",
-    "[Док] Паранормальные явления",
-    "[Док] Альтернативная история и наука",
-    "[Док] Внежанровая документалистика",
-    "[Док] Разное / некондиция"
-  ],
-  "Документальные (HD Video)": [
-    "Информационно-аналитические и общественно-политические (HD Video)",
-    "Биографии. Личности и кумиры (HD Video)",
-    "Военное дело (HD Video)",
-    "Естествознание, наука и техника (HD Video)",
-    "Путешествия и туризм (HD Video)",
-    "Флора и фауна (HD Video)",
-    "История (HD Video)",
-    "BBC, Discovery, National Geographic (HD Video)",
-    "Криминальная документалистика (HD Video)"
-  ],
-  "Развлекательные телепередачи и шоу, приколы и юмор": [
-    "[Видео Юмор] Интеллектуальные игры и викторины",
-    "[Видео Юмор] Реалити и ток-шоу / номинации / показы",
-    "[Видео Юмор] Детские телешоу",
-    "[Видео Юмор] КВН",
-    "[Видео Юмор] Пост КВН",
-    "[Видео Юмор] Кривое Зеркало / Городок / В Городке",
-    "[Видео Юмор] Ледовые шоу",
-    "[Видео Юмор] Музыкальные шоу",
-    "[Видео Юмор] Званый ужин",
-    "[Видео Юмор] Хорошие Шутки",
-    "[Видео Юмор] Вечерний Квартал",
-    "[Видео Юмор] Фильмы со смешным переводом (пародии)",
-    "[Видео Юмор] Stand-up comedy",
-    "[Видео Юмор] Украинские Шоу",
-    "[Видео Юмор] Танцевальные шоу, концерты, выступления",
-    "[Видео Юмор] Цирк",
-    "[Видео Юмор] Школа злословия",
-    "[Видео Юмор] Сатирики и юмористы",
-    "Юмористические аудиопередачи",
-    "Аудио и видео ролики (Приколы и юмор)"
-  ],
-  "Зимние Олимпийские игры 2018": [
-    "Биатлон",
-    "Лыжные гонки",
-    "Прыжки на лыжах с трамплина / Лыжное двоеборье",
-    "Горные лыжи / Сноубординг / Фристайл",
-    "Бобслей / Санный спорт / Скелетон",
-    "Конькобежный спорт / Шорт-трек",
-    "Фигурное катание",
-    "Хоккей",
-    "Керлинг",
-    "Обзорные и аналитические программы"
-  ],
-  "Спортивные турниры, фильмы и передачи": [
-    "Автоспорт",
-    "Мотоспорт",
-    "Формула-1 (2018)",
-    "Формула-1 (2012-2017)",
-    "Формула 1 (до 2011 вкл.)",
-    "Велоспорт",
-    "Волейбол/Гандбол",
-    "Бильярд",
-    "Покер",
-    "Бодибилдинг/Силовые виды спорта",
-    "Бокс",
-    "Классические единоборства",
-    "Смешанные единоборства и K-1",
-    "Американский футбол",
-    "Регби",
-    "Бейсбол",
-    "Теннис",
-    "Бадминтон/Настольный теннис",
-    "Гимнастика/Соревнования по танцам",
-    "Лёгкая атлетика/Водные виды спорта",
-    "Зимние виды спорта",
-    "Фигурное катание",
-    "Биатлон",
-    "Экстрим",
-    "Спорт (видео)"
-  ],
-  "⚽ Футбол": [
-    "Чемпионат Мира 2018 (плей-офф финального турнира)",
-    "Чемпионат Мира 2018 (групповой этап финального турнира)",
-    "Чемпионат Мира 2018 (обзорные передачи, документалистика)",
-    "Чемпионат Мира 2018 (отборочный турнир)",
-    "Чемпионаты Мира",
-    "Россия 2018-2019",
-    "Лига Наций 2018",
-    "Товарищеские турниры и матчи",
-    "Россия/СССР",
-    "Англия",
-    "Испания",
-    "Италия",
-    "Германия",
-    "Франция",
-    "Украина",
-    "Другие национальные чемпионаты и кубки",
-    "Международные турниры",
-    "Еврокубки 2018-2019",
-    "Еврокубки 2017-2018",
-    "Еврокубки 2011-2017",
-    "Еврокубки",
-    "Чемпионаты Европы",
-    "Обзорные и аналитические передачи 2018-2019",
-    "Обзорные и аналитические передачи",
-    "Мини-футбол/Пляжный футбол"
-  ],
-  "🏀 Баскетбол": [
-    "Международные соревнования",
-    "NBA / NCAA (до 2000 г.)",
-    "NBA / NCAA (2000-2010 гг.)",
-    "NBA / NCAA (2010-2019 гг.)",
-    "Европейский клубный баскетбол"
-  ],
-  "🏒 Хоккей": [
-    "Хоккей с мячом / Бенди",
-    "Международные турниры",
-    "КХЛ",
-    "НХЛ (до 2011/12)",
-    "НХЛ (с 2013)",
-    "СССР - Канада",
-    "Документальные фильмы и аналитика"
-  ],
-  "Рестлинг": [
-    "Professional Wrestling",
-    "Independent Wrestling",
-    "International Wrestling",
-    "Oldschool Wrestling",
-    "Documentary Wrestling"
-  ],
-  "Сканирование, обработка сканов": [],
-  "Книги и журналы (общий раздел)": [
-    "Кино, театр, ТВ, мультипликация, цирк",
-    "Рисунок, графический дизайн",
-    "Фото и видеосъемка",
-    "Журналы и газеты (общий раздел)",
-    "Эзотерика, гадания, магия, фен-шуй",
-    "Астрология",
-    "Красота. Уход. Домоводство",
-    "Мода. Стиль. Этикет",
-    "Путешествия и туризм",
-    "Знаменитости и кумиры",
-    "Разное (книги)"
-  ],
-  "Для детей, родителей и учителей": [
-    "Учебная литература для детского сада и начальной школы (до 4 класс..",
-    "Учебная литература для старших классов (5-11 класс)",
-    "Учителям и педагогам",
-    "Научно-популярная и познавательная литература (для детей)",
-    "Досуг и творчество",
-    "Воспитание и развитие",
-    "Худ. лит-ра для дошкольников и младших классов",
-    "Худ. лит-ра для средних и старших классов"
-  ],
-  "Спорт, физическая культура, боевые искусства": [
-    "Футбол (книги и журналы)",
-    "Хоккей (книги и журналы)",
-    "Игровые виды спорта",
-    "Легкая атлетика. Плавание. Гимнастика. Тяжелая атлетика. Гребля",
-    "Автоспорт. Мотоспорт. Велоспорт",
-    "Шахматы. Шашки",
-    "Боевые искусства, единоборства",
-    "Экстрим (книги)",
-    "Физкультура, фитнес, бодибилдинг",
-    "Спортивная пресса"
-  ],
-  "Гуманитарные науки": [
-    "Искусствоведение. Культурология",
-    "Фольклор. Эпос. Мифология",
-    "Литературоведение",
-    "Лингвистика",
-    "Философия",
-    "Политология",
-    "Социология",
-    "Публицистика, журналистика",
-    "Бизнес, менеджмент",
-    "Маркетинг",
-    "Экономика",
-    "Финансы",
-    "Юридические науки. Право. Криминалистика"
-  ],
-  "Исторические науки": [
-    "Методология и философия исторической науки",
-    "Исторические источники (книги, периодика)",
-    "Исторические источники (документы)",
-    "Исторические персоны",
-    "Альтернативные исторические теории",
-    "Археология",
-    "Древний мир. Античность",
-    "Средние века",
-    "История Нового и Новейшего времени",
-    "История Европы",
-    "История Азии и Африки",
-    "История Америки, Австралии, Океании",
-    "История России",
-    "Эпоха СССР",
-    "История стран бывшего СССР",
-    "Этнография, антропология",
-    "Международные отношения. Дипломатия"
-  ],
-  "Точные, естественные и инженерные науки": [
-    "Авиация / Космонавтика",
-    "Физика",
-    "Астрономия",
-    "Биология / Экология",
-    "Химия / Биохимия",
-    "Математика",
-    "География / Геология / Геодезия",
-    "Электроника / Радио",
-    "Схемы и сервис-мануалы (оригинальная документация)",
-    "Архитектура / Строительство / Инженерные сети",
-    "Машиностроение",
-    "Сварка / Пайка / Неразрушающий контроль",
-    "Автоматизация / Робототехника",
-    "Металлургия / Материаловедение",
-    "Механика, сопротивление материалов",
-    "Энергетика / электротехника",
-    "Нефтяная, газовая и химическая промышленность",
-    "Сельское хозяйство и пищевая промышленность",
-    "Железнодорожное дело",
-    "Нормативная документация",
-    "Журналы: научные, научно-популярные, радио и др."
-  ],
-  "Ноты и Музыкальная литература": [
-    "Академическая музыка (Ноты и Media CD)",
-    "Другие направления (Ноты, табулатуры)",
-    "Самоучители и Школы",
-    "Песенники (Songbooks)",
-    "Музыкальная литература и Теория",
-    "Музыкальные журналы"
-  ],
-  "Военное дело": [
-    "Милитария",
-    "Военная история",
-    "История Второй мировой войны",
-    "Биографии и мемуары военных деятелей",
-    "Военная техника",
-    "Стрелковое оружие",
-    "Учебно-методическая литература",
-    "Спецслужбы мира"
-  ],
-  "Психология": [
-    "Общая и прикладная психология",
-    "Психотерапия и консультирование",
-    "Психодиагностика и психокоррекция",
-    "Социальная психология и психология отношений",
-    "Тренинг и коучинг",
-    "Саморазвитие и самосовершенствование",
-    "Популярная психология",
-    "Сексология. Взаимоотношения полов (18+)"
-  ],
-  "Коллекционирование, увлечения и хобби": [
-    "Коллекционирование и вспомогательные ист. дисциплины",
-    "Вышивание",
-    "Вязание",
-    "Шитье, пэчворк",
-    "Кружевоплетение",
-    "Бисероплетение. Ювелирика. Украшения из проволоки.",
-    "Бумажный арт",
-    "Другие виды декоративно-прикладного искусства",
-    "Домашние питомцы и аквариумистика",
-    "Охота и рыбалка",
-    "Кулинария (книги)",
-    "Кулинария (газеты и журналы)",
-    "Моделизм",
-    "Приусадебное хозяйство / Цветоводство",
-    "Ремонт, частное строительство, дизайн интерьеров",
-    "Деревообработка",
-    "Настольные игры",
-    "Прочие хобби"
-  ],
-  "Художественная литература": [
-    "Русская литература",
-    "Зарубежная литература (до 1900 г.)",
-    "Зарубежная литература (XX и XXI век)",
-    "Детектив, боевик",
-    "Женский роман",
-    "Отечественная фантастика / фэнтези / мистика",
-    "Зарубежная фантастика / фэнтези / мистика",
-    "Приключения",
-    "Литературные журналы"
-  ],
-  "Компьютерная литература": [
-    "Программы от Microsoft",
-    "Другие программы",
-    "Mac OS; Linux, FreeBSD и прочие *NIX",
-    "СУБД",
-    "Веб-дизайн и программирование",
-    "Программирование (книги)",
-    "Графика, обработка видео",
-    "Сети / VoIP",
-    "Хакинг и безопасность",
-    "Железо (книги о ПК)",
-    "Инженерные и научные программы (книги)",
-    "Компьютерные журналы и приложения к ним",
-    "Дисковые приложения к игровым журналам"
-  ],
-  "Комиксы, манга, ранобэ": [
-    "Комиксы на русском языке",
-    "Комиксы издательства Marvel",
-    "Комиксы издательства DC",
-    "Комиксы других издательств",
-    "Комиксы на других языках",
-    "Манга (на русском языке)",
-    "Манга (на иностранных языках)",
-    "Ранобэ"
-  ],
-  "Коллекции книг и библиотеки": [
-    "Библиотеки (зеркала сетевых библиотек/коллекций)",
-    "Тематические коллекции (подборки)",
-    "Многопредметные коллекции (подборки)"
-  ],
-  "Мультимедийные и интерактивные издания": [
-    "Мультимедийные энциклопедии",
-    "Интерактивные обучающие и развивающие материалы",
-    "Обучающие издания для детей",
-    "Кулинария. Цветоводство. Домоводство",
-    "Культура. Искусство. История"
-  ],
-  "Медицина и здоровье": [
-    "Клиническая медицина до 1980 г.",
-    "Клиническая медицина с 1980 по 2000 г.",
-    "Клиническая медицина после 2000 г.",
-    "Научная медицинская периодика (газеты и журналы)",
-    "Медико-биологические науки",
-    "Фармация и фармакология",
-    "Популярная медицинская периодика (газеты и журналы)",
-    "Нетрадиционная, народная медицина и популярные книги о здоровье",
-    "Ветеринария, разное",
-    "Тематические коллекции книг"
-  ],
-  "Иностранные языки для взрослых": [
-    "Английский язык (для взрослых)",
-    "Немецкий язык",
-    "Французский язык",
-    "Испанский язык",
-    "Итальянский язык",
-    "Финский язык",
-    "Другие европейские языки",
-    "Арабский язык",
-    "Китайский язык",
-    "Японский язык",
-    "Другие восточные языки",
-    "Русский язык как иностранный",
-    "Мультиязычные сборники",
-    "LIM-курсы",
-    "Разное (иностранные языки)"
-  ],
-  "Иностранные языки для детей": [
-    "Английский язык (для детей)",
-    "Другие европейские языки (для детей)",
-    "Восточные языки (для детей)",
-    "Школьные учебники, ЕГЭ"
-  ],
-  "Художественная литература (ин.языки)": [
-    "Художественная литература на английском языке",
-    "Художественная литература на французском языке",
-    "Художественная литература на других европейских языках",
-    "Художественная литература на восточных языках"
-  ],
-  "Аудиокниги на иностранных языках": [
-    "Аудиокниги на английском языке",
-    "Аудиокниги на немецком языке",
-    "Аудиокниги на других иностранных языках"
-  ],
-  "Видеоуроки и обучающие интерактивные DVD": [
-    "Кулинария",
-    "Спорт",
-    "Фитнес - Кардио-Силовые Тренировки",
-    "Фитнес - Разум и Тело",
-    "Бодибилдинг",
-    "Оздоровительные практики",
-    "Йога",
-    "Видео- и фотосъёмка",
-    "Уход за собой",
-    "Рисование",
-    "Игра на гитаре",
-    "Ударные инструменты",
-    "Другие музыкальные инструменты",
-    "Игра на бас-гитаре",
-    "Бальные танцы",
-    "Танец живота",
-    "Уличные и клубные танцы",
-    "Танцы, разное",
-    "Охота",
-    "Рыболовство и подводная охота",
-    "Фокусы и трюки",
-    "Образование",
-    "Финансы",
-    "Продажи, бизнес",
-    "Беременность, роды, материнство",
-    "Учебные видео для детей",
-    "Психология",
-    "Эзотерика, саморазвитие",
-    "Пикап, знакомства",
-    "Строительство, ремонт и дизайн",
-    "Дерево- и металлообработка",
-    "Растения и животные",
-    "Хобби и рукоделие",
-    "Медицина и стоматология",
-    "Психотерапия и клиническая психология",
-    "Массаж",
-    "Здоровье",
-    "Медицина - интерактивный софт",
-    "Разное"
-  ],
-  "Боевые искусства (Видеоуроки)": [
-    "Айкидо и айки-дзюцу",
-    "Вин чун",
-    "Джиу-джитсу",
-    "Дзюдо и самбо",
-    "Каратэ",
-    "Работа с оружием",
-    "Русский стиль",
-    "Рукопашный бой",
-    "Смешанные стили",
-    "Ударные стили",
-    "Ушу",
-    "Разное"
-  ],
-  "Компьютерные видеоуроки и обучающие интерактивные DVD": [
-    "Компьютерные сети и безопасность",
-    "ОС и серверные программы Microsoft",
-    "Офисные программы Microsoft",
-    "ОС и программы семейства UNIX",
-    "Adobe Photoshop",
-    "Autodesk Maya",
-    "Autodesk 3ds Max",
-    "Autodesk Softimage (XSI)",
-    "ZBrush",
-    "Flash, Flex и ActionScript",
-    "2D-графика",
-    "3D-графика",
-    "Инженерные и научные программы (видеоуроки)",
-    "Web-дизайн",
-    "WEB, SMM, SEO, интернет-маркетинг",
-    "Программирование (видеоуроки)",
-    "Программы для Mac OS",
-    "Работа с видео",
-    "Работа со звуком",
-    "Разное (Компьютерные видеоуроки)"
-  ],
-  "Радиоспектакли, история, мемуары": [
-    "[Аудио] Радиоспектакли и литературные чтения",
-    "[Аудио] Жизнь замечательных людей",
-    "[Аудио] История, культурология, философия"
-  ],
-  "Фантастика, фэнтези, мистика, ужасы, фанфики": [
-    "[Аудио] Зарубежная фантастика, фэнтези, мистика, ужасы, фанфики",
-    "[Аудио] Российская фантастика, фэнтези, мистика, ужасы, фанфики",
-    "[Аудио] Любовно-фантастический роман",
-    "[Аудио] Сборники/разное Фантастика, фэнтези, мистика, ужасы, фанфи.."
-  ],
-  "Религии": [
-    "[Аудио] Православие",
-    "[Аудио] Ислам",
-    "[Аудио] Другие традиционные религии",
-    "[Аудио] Нетрадиционные религиозно-философские учения"
-  ],
-  "Прочая литература": [
-    "[Аудио] Книги по медицине",
-    "[Аудио] Учебная и научно-популярная литература",
-    "[Аудио] lossless-аудиокниги",
-    "[Аудио] Бизнес",
-    "[Аудио] Разное",
-    "[Аудио] Некондиционные раздачи"
-  ],
-  "Ремонт и эксплуатация транспортных средств": [
-    "Оригинальные каталоги по подбору запчастей",
-    "Неоригинальные каталоги по подбору запчастей",
-    "Программы по диагностике и ремонту",
-    "Тюнинг, чиптюнинг, настройка",
-    "Книги по ремонту/обслуживанию/эксплуатации ТС",
-    "Мультимедийки по ремонту/обслуживанию/эксплуатации ТС",
-    "Учет, утилиты и прочее",
-    "Виртуальная автошкола",
-    "Видеоуроки по вождению транспортных средств",
-    "Видеоуроки по ремонту транспортных средств",
-    "Журналы по авто/мото",
-    "Водный транспорт"
-  ],
-  "Фильмы и передачи по авто/мото": [
-    "Документальные/познавательные фильмы",
-    "Развлекательные передачи",
-    "Top Gear/Топ Гир",
-    "Тест драйв/Обзоры/Автосалоны",
-    "Тюнинг/форсаж"
-  ],
-  "Классическая и современная академическая музыка": [
-    "Классическая музыка (Видео)",
-    "Классическая музыка (DVD и HD Видео)",
-    "Опера (Видео)",
-    "Опера (DVD и HD Видео)",
-    "Балет и современная хореография (Видео, DVD и HD Видео)",
-    "Полные собрания сочинений и многодисковые издания (lossless)",
-    "Опера (lossless)",
-    "Вокальная музыка (lossless)",
-    "Хоровая музыка (lossless)",
-    "Оркестровая музыка (lossless)",
-    "Концерт для инструмента с оркестром (lossless)",
-    "Камерная инструментальная музыка (lossless)",
-    "Сольная инструментальная музыка (lossless)",
-    "Духовные песнопения и музыка (lossless)",
-    "Духовные песнопения и музыка (lossy)",
-    "Полные собрания сочинений и многодисковые издания (lossy)",
-    "Вокальная и хоровая музыка (lossy)",
-    "Оркестровая музыка (lossy)",
-    "Камерная и сольная инструментальная музыка (lossy)",
-    "Классика в современной обработке, Classical Crossover (lossy и los.."
-  ],
-  "Фольклор, Народная и Этническая музыка": [
-    "Восточноевропейский фолк (lossy)",
-    "Восточноевропейский фолк (lossless)",
-    "Западноевропейский фолк (lossy)",
-    "Западноевропейский фолк (lossless)",
-    "Klezmer и Еврейский фольклор (lossy и lossless)",
-    "Этническая музыка Сибири, Средней и Восточной Азии (lossy)",
-    "Этническая музыка Сибири, Средней и Восточной Азии (lossless)",
-    "Этническая музыка Индии (lossy)",
-    "Этническая музыка Индии (lossless)",
-    "Этническая музыка Африки и Ближнего Востока (lossy)",
-    "Этническая музыка Африки и Ближнего Востока (lossless)",
-    "Фольклорная, Народная, Эстрадная музыка Кавказа и Закавказья (loss..",
-    "Этническая музыка Северной и Южной Америки (lossy)",
-    "Этническая музыка Северной и Южной Америки (lossless)",
-    "Этническая музыка Австралии, Тихого и Индийского океанов (lossy и ..",
-    "Country, Bluegrass (lossy)",
-    "Country, Bluegrass (lossless)",
-    "Фольклор, Народная и Этническая музыка (Видео)",
-    "Фольклор, Народная и Этническая музыка (DVD Video)",
-    "Фольклор, Народная и Этническая музыка (HD Видео)"
-  ],
-  "New Age, Relax, Meditative & Flamenco": [
-    "New Age & Meditative (lossy)",
-    "New Age & Meditative (lossless)",
-    "Фламенко и акустическая гитара (lossy)",
-    "Фламенко и акустическая гитара (lossless)",
-    "Музыка для бальных танцев (lossy и lossless)",
-    "New Age, Relax, Meditative & Flamenco (Видео)",
-    "New Age, Relax, Meditative & Flamenco (DVD и HD Видео)",
-    "Звуки природы"
-  ],
-  "Рэп, Хип-Хоп, R'n'B": [
-    "Отечественный Рэп, Хип-Хоп (lossy)",
-    "Отечественный R'n'B (lossy)",
-    "Отечественный Рэп, Хип-Хоп, R'n'B (lossless)",
-    "Зарубежный R'n'B (lossy)",
-    "Зарубежный Рэп, Хип-Хоп (lossy)",
-    "Зарубежный Рэп, Хип-Хоп (lossless)",
-    "Зарубежный R'n'B (lossless)",
-    "Отечественный Рэп, Хип-Хоп (Видео)",
-    "Отечественный R'n'B (Видео)",
-    "Зарубежный Рэп, Хип-Хоп (Видео)",
-    "Зарубежный R'n'B (Видео)",
-    "Рэп, Хип-Хоп, R'n'B (DVD Video)",
-    "Рэп, Хип-Хоп, R'n'B (HD Видео)"
-  ],
-  "Reggae, Ska, Dub": [
-    "Rocksteady, Early Reggae, Ska-Jazz, Trad.Ska (lossy и lossless)",
-    "Punky-Reggae, Rocksteady-Punk, Ska Revival (lossy)",
-    "3rd Wave Ska (lossy)",
-    "Ska-Punk, Ska-Core (lossy)",
-    "Reggae (lossy)",
-    "Dub (lossy)",
-    "Dancehall, Raggamuffin (lossy)",
-    "Reggae, Dancehall, Dub (lossless)",
-    "Ska, Ska-Punk, Ska-Jazz (lossless)",
-    "Отечественный Reggae, Dub (lossy и lossless)",
-    "Отечественная Ska музыка (lossy и lossless)",
-    "Reggae, Ska, Dub (компиляции) (lossy и lossless)",
-    "Reggae, Ska, Dub (Видео)",
-    "Reggae, Ska, Dub (DVD и HD Video)"
-  ],
-  "Саундтреки, караоке и мюзиклы": [
-    "Караоке (аудио)",
-    "Караоке (видео)",
-    "Минусовки (lossy и lossless)",
-    "Саундтреки к отечественным фильмам (lossless)",
-    "Саундтреки к отечественным фильмам (lossy)",
-    "Саундтреки к зарубежным фильмам (lossless)",
-    "Саундтреки к зарубежным фильмам (lossy)",
-    "Саундтреки к сериалам (lossless)",
-    "Саундтреки к сериалам (lossy)",
-    "Саундтреки к мультфильмам (lossy и lossless)",
-    "Саундтреки к аниме (lossless)",
-    "Саундтреки к аниме (lossy)",
-    "Неофициальные саундтреки к фильмам и сериалам (lossy)",
-    "Саундтреки к играм (lossless)",
-    "Саундтреки к играм (lossy)",
-    "Неофициальные саундтреки к играм (lossy)",
-    "Аранжировки музыки из игр (lossy и lossless)",
-    "Мюзикл (lossy и lossless)",
-    "Мюзикл (Видео и DVD Video)"
-  ],
-  "Шансон, Авторская и Военная песня": [
-    "Отечественный шансон (lossless)",
-    "Отечественный шансон (lossy)",
-    "Сборники отечественного шансона (lossy)",
-    "Военная песня, марши (lossless)",
-    "Военная песня, марши (lossy)",
-    "Авторская песня (lossless)",
-    "Авторская песня (lossy)",
-    "Менестрели и ролевики (lossy и lossless)",
-    "Видео (Шансон, Авторская и Военная песня)",
-    "DVD Видео (Шансон, Авторская и Военная песня)"
-  ],
-  "Музыка других жанров": [
-    "Видео (Музыка других жанров)",
-    "DVD Video (Музыка других жанров)"
-  ],
-  "Отечественная поп-музыка": [
-    "Отечественная поп-музыка (lossy)",
-    "Отечественная поп-музыка (сборники) (lossy)",
-    "Отечественная поп-музыка (lossless)",
-    "Советская эстрада, ретро, романсы (lossy)",
-    "Советская эстрада, ретро, романсы (lossless)"
-  ],
-  "Зарубежная поп-музыка": [
-    "Зарубежная поп-музыка (lossy)",
-    "Зарубежная поп-музыка (сборники) (lossy)",
-    "Зарубежная поп-музыка (lossless)",
-    "Итальянская поп-музыка (lossy)",
-    "Итальянская поп-музыка (lossless)",
-    "Латиноамериканская поп-музыка (lossy)",
-    "Латиноамериканская поп-музыка (lossless)",
-    "Восточноазиатская поп-музыка (lossy)",
-    "Восточноазиатская поп-музыка (lossless)",
-    "Зарубежный шансон (lossy)",
-    "Зарубежный шансон (lossless)",
-    "Easy Listening, Instrumental Pop (lossy)",
-    "Easy Listening, Instrumental Pop (lossless)",
-    "Сборники песен для детей (lossy и lossless)"
-  ],
-  "Eurodance, Disco, Hi-NRG": [
-    "Eurodance, Euro-House, Technopop (lossy)",
-    "Eurodance, Euro-House, Technopop (сборники) (lossy)",
-    "Eurodance, Euro-House, Technopop (lossless)",
-    "Disco, Italo-Disco, Euro-Disco, Hi-NRG (lossy)",
-    "Disco, Italo-Disco, Euro-Disco, Hi-NRG (сборники) (lossy)",
-    "Disco, Italo-Disco, Euro-Disco, Hi-NRG (lossless)"
-  ],
-  "Видео, DVD Video, HD Video (поп-музыка)": [
-    "Отечественная поп-музыка (Видео)",
-    "Отечественная поп-музыка (DVD Video)",
-    "Советская эстрада, ретро, романсы (Видео)",
-    "Советская эстрада, ретро, романсы (DVD Video)",
-    "Зарубежная поп-музыка (Видео)",
-    "Зарубежная поп-музыка (DVD Video)",
-    "Eurodance, Disco (Видео)",
-    "Eurodance, Disco (DVD Video)",
-    "Восточноазиатская поп-музыка (Видео)",
-    "Восточноазиатская поп-музыка (DVD Video)",
-    "Зарубежный шансон (Видео)",
-    "Зарубежный шансон (DVD Video)",
-    "Отечественная поп-музыка (Сборные концерты, док. видео) (Видео и D..",
-    "Зарубежная поп-музыка (Сборные концерты, док. видео) (Видео и DVD)",
-    "Отечественная Поп-музыка, Шансон, Eurodance, Disco (HD Video)",
-    "Зарубежная Поп-музыка, Шансон, Eurodance, Disco (HD Video)"
-  ],
-  "Зарубежный джаз": [
-    "Early Jazz, Swing, Gypsy (lossless)",
-    "Bop (lossless)",
-    "Mainstream Jazz, Cool (lossless)",
-    "Jazz Fusion (lossless)",
-    "World Fusion, Ethnic Jazz (lossless)",
-    "Avant-Garde Jazz, Free Improvisation (lossless)",
-    "Modern Creative, Third Stream (lossless)",
-    "Smooth, Jazz-Pop (lossless)",
-    "Vocal Jazz (lossless)",
-    "Funk, Soul, R&B (lossless)",
-    "Сборники зарубежного джаза (lossless)",
-    "Зарубежный джаз (lossy)"
-  ],
-  "Зарубежный блюз": [
-    "Blues (Texas, Chicago, Modern and Others) (lossless)",
-    "Blues-rock (lossless)",
-    "Roots, Pre-War Blues, Early R&B, Gospel (lossless)",
-    "Зарубежный блюз (сборники; Tribute VA) (lossless)",
-    "Зарубежный блюз (lossy)"
-  ],
-  "Отечественный джаз и блюз": [
-    "Отечественный джаз (lossless)",
-    "Отечественный джаз (lossy)",
-    "Отечественный блюз (lossless)",
-    "Отечественный блюз (lossy)"
-  ],
-  "Видео, DVD Video, HD Video (Джаз и блюз)": [
-    "Джаз и Блюз (Видео)",
-    "Джаз и Блюз (DVD Видео)",
-    "Джаз и Блюз (HD Video)"
-  ],
-  "Зарубежный Rock": [
-    "Classic Rock & Hard Rock (lossless)",
-    "Classic Rock & Hard Rock (lossy)",
-    "Progressive & Art-Rock (lossless)",
-    "Progressive & Art-Rock (lossy)",
-    "Folk-Rock (lossless)",
-    "Folk-Rock (lossy)",
-    "AOR (Melodic Hard Rock, Arena rock) (lossless)",
-    "AOR (Melodic Hard Rock, Arena rock) (lossy)",
-    "Pop-Rock & Soft Rock (lossless)",
-    "Pop-Rock & Soft Rock (lossy)",
-    "Instrumental Guitar Rock (lossless)",
-    "Instrumental Guitar Rock (lossy)",
-    "Rockabilly, Psychobilly, Rock'n'Roll (lossless)",
-    "Rockabilly, Psychobilly, Rock'n'Roll (lossy)",
-    "Сборники зарубежного рока (lossless)",
-    "Сборники зарубежного рока (lossy)",
-    "Восточноазиатский рок (lossless)",
-    "Восточноазиатский рок (lossy)"
-  ],
-  "Зарубежный Metal": [
-    "Avant-garde, Experimental Metal (lossless)",
-    "Avant-garde, Experimental Metal (lossy)",
-    "Black (lossless)",
-    "Black (lossy)",
-    "Death, Doom (lossless)",
-    "Death, Doom (lossy)",
-    "Folk, Pagan, Viking (lossless)",
-    "Folk, Pagan, Viking (lossy)",
-    "Gothic Metal (lossless)",
-    "Gothic Metal (lossy)",
-    "Grind, Brutal Death (lossless)",
-    "Grind, Brutal Death (lossy)",
-    "Heavy, Power, Progressive (lossless)",
-    "Heavy, Power, Progressive (lossy)",
-    "Sludge, Stoner, Post-Metal (lossless)",
-    "Sludge, Stoner, Post-Metal (lossy)",
-    "Thrash, Speed (lossless)",
-    "Thrash, Speed (lossy)",
-    "Сборники (lossless)",
-    "Сборники (lossy)"
-  ],
-  "Зарубежные Alternative, Punk, Independent": [
-    "Alternative & Nu-metal (lossless)",
-    "Alternative & Nu-metal (lossy)",
-    "Punk (lossless)",
-    "Punk (lossy)",
-    "Hardcore (lossless)",
-    "Hardcore (lossy)",
-    "Indie, Post-Rock & Post-Punk (lossless)",
-    "Indie, Post-Rock & Post-Punk (lossy)",
-    "Industrial & Post-industrial (lossless)",
-    "Industrial & Post-industrial (lossy)",
-    "Emocore, Post-hardcore, Metalcore (lossless)",
-    "Emocore, Post-hardcore, Metalcore (lossy)",
-    "Gothic Rock & Dark Folk (lossless)",
-    "Gothic Rock & Dark Folk (lossy)",
-    "Avant-garde, Experimental Rock (lossless)",
-    "Avant-garde, Experimental Rock (lossy)"
-  ],
-  "Отечественный Rock, Metal": [
-    "Rock (lossless)",
-    "Rock (lossy)",
-    "Alternative, Punk, Independent (lossless)",
-    "Alternative, Punk, Independent (lossy)",
-    "Metal (lossless)",
-    "Metal (lossy)",
-    "Rock на языках народов xUSSR (lossless)",
-    "Rock на языках народов xUSSR (lossy)"
-  ],
-  "Видео, DVD Video, HD Video (Рок-музыка)": [
-    "Rock (Видео)",
-    "Rock (DVD Video)",
-    "Rock (Неофициальные DVD Video)",
-    "Metal (Видео)",
-    "Metal (DVD Video)",
-    "Metal (Неофициальные DVD Video)",
-    "Alternative, Punk, Independent (Видео)",
-    "Alternative, Punk, Independent (DVD Video)",
-    "Alternative, Punk, Independent (Неофициальные DVD Video)",
-    "Отечественный Рок, Панк, Альтернатива (Видео)",
-    "Отечественный Рок, Панк, Альтернатива (DVD Video)",
-    "Отечественный Металл (Видео)",
-    "Отечественный Металл (DVD Video)",
-    "Отечественный Рок, Панк, Альтернатива, Металл (Неофициальные DVD V..",
-    "Рок-музыка (HD Video)"
-  ],
-  "Trance, Goa Trance, Psy-Trance, PsyChill, Ambient, Dub": [
-    "Goa Trance, Psy-Trance (lossless)",
-    "Goa Trance, Psy-Trance (lossy)",
-    "PsyChill, Ambient, Dub (lossless)",
-    "PsyChill, Ambient, Dub (lossy)",
-    "Goa Trance, Psy-Trance, PsyChill, Ambient, Dub (Live Sets, Mixes) ..",
-    "Trance (lossless)",
-    "Trance (lossy)",
-    "Trance (Singles, EPs) (lossy)",
-    "Trance (Radioshows, Podcasts, Live Sets, Mixes) (lossy)"
-  ],
-  "House, Techno, Hardcore, Hardstyle, Jumpstyle": [
-    "Hardcore, Hardstyle, Jumpstyle (lossless)",
-    "Hardcore, Hardstyle, Jumpstyle (lossy)",
-    "Hardcore, Hardstyle, Jumpstyle (vinyl, web)",
-    "House (lossless)",
-    "House (Radioshow, Podcast, Liveset, Mixes)",
-    "House (lossy)",
-    "House (Проморелизы, сборники) (lossy)",
-    "House (Singles, EPs) (lossy)",
-    "Techno (lossless)",
-    "Techno (lossy)",
-    "Techno (Radioshows, Podcasts, Livesets, Mixes)",
-    "Techno (Singles, EPs) (lossy)"
-  ],
-  "Drum & Bass, Jungle, Breakbeat, Dubstep, IDM, Electro": [
-    "Electro, Electro-Freestyle, Nu Electro (lossless)",
-    "Electro, Electro-Freestyle, Nu Electro (lossy)",
-    "Drum & Bass, Jungle (lossless)",
-    "Drum & Bass, Jungle (lossy)",
-    "Drum & Bass, Jungle (Radioshows, Podcasts, Livesets, Mixes)",
-    "Breakbeat (lossless)",
-    "Breakbeat (lossy)",
-    "Dubstep (lossless)",
-    "Dubstep (lossy)",
-    "Breakbeat, Dubstep (Radioshows, Podcasts, Livesets, Mixes)",
-    "IDM (lossless)",
-    "IDM (lossy)",
-    "IDM Discography & Collections (lossy)"
-  ],
-  "Chillout, Lounge, Downtempo, Trip-Hop": [
-    "Chillout, Lounge, Downtempo (lossless)",
-    "Chillout, Lounge, Downtempo (lossy)",
-    "Nu Jazz, Acid Jazz, Future Jazz (lossless)",
-    "Nu Jazz, Acid Jazz, Future Jazz (lossy)",
-    "Trip Hop, Abstract Hip-Hop (lossless)",
-    "Trip Hop, Abstract Hip-Hop (lossy)"
-  ],
-  "Traditional Electronic, Ambient, Modern Classical, Electroacoustic, Ex..": [
-    "Traditional Electronic, Ambient (lossless)",
-    "Traditional Electronic, Ambient (lossy)",
-    "Modern Classical, Electroacoustic (lossless)",
-    "Modern Classical, Electroacoustic (lossy)",
-    "Experimental (lossless)",
-    "Experimental (lossy)",
-    "8-bit, Chiptune (lossy & lossless)"
-  ],
-  "Industrial, Noise, EBM, Dark Electro, Aggrotech, Synthpop, New Wave": [
-    "EBM, Dark Electro, Aggrotech (lossless)",
-    "EBM, Dark Electro, Aggrotech (lossy)",
-    "Industrial, Noise (lossless)",
-    "Industrial, Noise (lossy)",
-    "Synthpop, Futurepop, New Wave, Electropop (lossless)",
-    "Synthpop, Futurepop, New Wave, Electropop (lossy)",
-    "Synthwave, Spacesynth, Dreamwave, Retrowave, Outrun (lossless)",
-    "Synthwave, Spacesynth, Dreamwave, Retrowave, Outrun (lossy)",
-    "Darkwave, Neoclassical, Ethereal, Dungeon Synth (lossless)",
-    "Darkwave, Neoclassical, Ethereal, Dungeon Synth (lossy)"
-  ],
-  "Label Packs (lossless)": [],
-  "Label packs, Scene packs (lossy)": [],
-  "Электронная музыка (Видео, DVD Video, HD Video)": [
-    "Электронная музыка (Официальные DVD Video)",
-    "Электронная музыка (Неофициальные, любительские DVD Video)",
-    "Электронная музыка (Видео)",
-    "Электронная музыка (HD Video)"
-  ],
-  "Hi-Res stereo и многоканальная музыка": [
-    "Классика и классика в современной обработке (Hi-Res stereo)",
-    "Классика и классика в современной обработке (многоканальная музыка..",
-    "New Age, Relax, Meditative & Flamenco (Hi-Res stereo и многоканаль..",
-    "Саундтреки (Hi-Res stereo и многоканальная музыка)",
-    "Музыка разных жанров (Hi-Res stereo и многоканальная музыка)",
-    "Поп-музыка (Hi-Res stereo)",
-    "Поп-музыка (многоканальная музыка)",
-    "Джаз и Блюз (Hi-Res stereo)",
-    "Джаз и Блюз (многоканальная музыка)",
-    "Рок-музыка (Hi-Res stereo)",
-    "Рок-музыка (многоканальная музыка)",
-    "Электронная музыка (Hi-Res stereo)",
-    "Электронная музыка (многоканальная музыка)"
-  ],
-  "Оцифровки с аналоговых носителей": [
-    "Классика и классика в современной обработке (оцифровки)",
-    "Фольклор, народная и этническая музыка (оцифровки)",
-    "Rap, Hip-Hop, R'n'B, Reggae, Ska, Dub (оцифровки)",
-    "Саундтреки и мюзиклы (оцифровки)",
-    "Шансон, авторские, военные песни и марши (оцифровки)",
-    "Музыка других жанров (оцифровки)",
-    "Зарубежная поп-музыка (оцифровки)",
-    "Советская эстрада, ретро, романсы (оцифровки)",
-    "Отечественная поп-музыка (оцифровки)",
-    "Инструментальная поп-музыка (оцифровки)",
-    "Джаз и блюз (оцифровки)",
-    "Зарубежная рок-музыка (оцифровки)",
-    "Отечественная рок-музыка (оцифровки)",
-    "Электронная музыка (оцифровки)"
-  ],
-  "Неофициальные конверсии цифровых форматов": [
-    "Конверсии Quadraphonic",
-    "Конверсии SACD",
-    "Конверсии Blu-Ray, ADVD и DVD-Audio",
-    "Апмиксы-Upmixes/Даунмиксы-Downmix"
-  ],
-  "Игры для Windows": [
-    "Горячие Новинки",
-    "Экшены от первого лица",
-    "Экшены от третьего лица",
-    "Хорроры",
-    "Аркады",
-    "Файтинги",
-    "Приключения и квесты",
-    "Квесты в стиле \"Поиск предметов\"",
-    "Визуальные новеллы",
-    "Для самых маленьких",
-    "Логические игры",
-    "Многопользовательские игры",
-    "Ролевые игры",
-    "Симуляторы",
-    "Стратегии в реальном времени",
-    "Пошаговые стратегии",
-    "Шахматы",
-    "IBM PC-несовместимые"
-  ],
-  "Прочее для Windows-игр": [
-    "Официальные патчи, моды, плагины, дополнения",
-    "Неофициальные моды, плагины, дополнения",
-    "Русификаторы"
-  ],
-  "Прочее для Microsoft Flight Simulator, Prepar3D, X-Plane": [
-    "Сценарии, меши и аэропорты для FS2004, FSX, P3D",
-    "Самолёты и вертолёты для FS2004, FSX, P3D",
-    "Миссии, трафик, звуки, паки и утилиты для FS2004, FSX, P3D",
-    "Сценарии, миссии, трафик, звуки, паки и утилиты для X-Plane",
-    "Самолёты и вертолёты для X-Plane"
-  ],
-  "Игры для Macintosh": [
-    "Нативные игры для Mac",
-    "Портированные игры для Mac"
-  ],
-  "Игры для Linux": [
-    "Нативные игры для Linux",
-    "Портированные игры для Linux"
-  ],
-  "Игры для консолей": [
-    "PS",
-    "PS2",
-    "PS3",
-    "Игры PS1, PS2 и PSP для PS3",
-    "PS4",
-    "PSP",
-    "Игры PS1 для PSP",
-    "PS Vita",
-    "Original Xbox",
-    "Xbox 360",
-    "Wii/WiiU",
-    "NDS/3DS",
-    "Switch",
-    "Dreamcast",
-    "Остальные платформы"
-  ],
-  "Видео для консолей": [
-    "Видео для PS Vita",
-    "Фильмы для PSP",
-    "Сериалы для PSP",
-    "Мультфильмы для PSP",
-    "Дорамы для PSP",
-    "Аниме для PSP",
-    "Видео для PSP",
-    "Видео для PS3 и других консолей"
-  ],
-  "Игры для мобильных устройств": [
-    "Игры для Android",
-    "Игры для Java",
-    "Игры для Symbian",
-    "Игры для Windows Mobile",
-    "Игры для Windows Phone"
-  ],
-  "Игровое видео": [
-    "Видеопрохождения игр"
-  ],
-  "Операционные системы от Microsoft": [
-    "Настольные ОС от Microsoft - Windows 8 и далее",
-    "Настольные ОС от Microsoft (выпущенные до Windows XP)",
-    "Настольные ОС от Microsoft - Windows XP - Windows 7",
-    "Серверные ОС от Microsoft",
-    "Разное (Операционные системы от Microsoft)"
-  ],
-  "Linux, Unix и другие ОС": [
-    "Операционные системы (Linux, Unix)",
-    "Программное обеспечение (Linux, Unix)",
-    "Другие ОС и ПО под них"
-  ],
-  "Тестовые диски для настройки аудио/видео аппаратуры": [],
-  "Системные программы": [
-    "Работа с жёстким диском",
-    "Резервное копирование",
-    "Архиваторы и файловые менеджеры",
-    "Программы для настройки и оптимизации ОС",
-    "Сервисное обслуживание компьютера",
-    "Работа с носителями информации",
-    "Информация и диагностика",
-    "Программы для интернет и сетей",
-    "ПО для защиты компьютера (Антивирусное ПО, Фаерволлы)",
-    "Анти-шпионы и анти-трояны",
-    "Программы для защиты информации",
-    "Драйверы и прошивки",
-    "Оригинальные диски к компьютерам и комплектующим",
-    "Серверное ПО для Windows",
-    "Изменение интерфейса ОС Windows",
-    "Скринсейверы",
-    "Разное (Системные программы под Windows)"
-  ],
-  "Системы для бизнеса, офиса, научной и проектной работы": [
-    "Всё для дома: кройка, шитьё, кулинария",
-    "Офисные системы",
-    "Системы для бизнеса",
-    "Распознавание текста, звука и синтез речи",
-    "Работа с PDF и DjVu",
-    "Словари, переводчики",
-    "Системы для научной работы",
-    "САПР (общие и машиностроительные)",
-    "САПР (электроника, автоматика, ГАП)",
-    "Программы для архитекторов и строителей",
-    "Библиотеки и проекты для архитекторов и дизайнеров интерьеров",
-    "Прочие справочные системы",
-    "Разное (Системы для бизнеса, офиса, научной и проектной работы)"
-  ],
-  "Веб-разработка и Программирование": [
-    "WYSIWYG Редакторы для веб-диза",
-    "Текстовые редакторы с подсветкой",
-    "Среды программирования, компиляторы и вспомогательные программы",
-    "Компоненты для сред программирования",
-    "Системы управления базами данных",
-    "Скрипты и движки сайтов, CMS а также расширения к ним",
-    "Шаблоны для сайтов и CMS",
-    "Разное (Веб-разработка и программирование)"
-  ],
-  "Программы для работы с мультимедиа и 3D": [
-    "Программные комплекты",
-    "Плагины для программ компании Adobe",
-    "Графические редакторы",
-    "Программы для верстки, печати и работы со шрифтами",
-    "3D моделирование, рендеринг и плагины для них",
-    "Анимация",
-    "Создание BD/HD/DVD-видео",
-    "Редакторы видео",
-    "Видео- Аудио- конверторы",
-    "Аудио- и видео-, CD- проигрыватели и каталогизаторы",
-    "Каталогизаторы и просмотрщики графики",
-    "Разное (Программы для работы с мультимедиа и 3D)",
-    "Виртуальные студии, секвенсоры и аудиоредакторы",
-    "Виртуальные инструменты и синтезаторы",
-    "Плагины для обработки звука",
-    "Разное (Программы для работы со звуком)"
-  ],
-  "Материалы для мультимедиа и дизайна": [
-    "Авторские работы",
-    "Официальные сборники векторных клипартов",
-    "Прочие векторные клипарты",
-    "Photostoсks",
-    "Костюмы для фотомонтажа",
-    "Рамки и виньетки для оформления фотографий",
-    "Прочие растровые клипарты",
-    "3D модели, сцены и материалы",
-    "Футажи",
-    "Прочие сборники футажей",
-    "Музыкальные библиотеки",
-    "Звуковые эффекты",
-    "Библиотеки сэмплов",
-    "Библиотеки и саундбанки для сэмплеров, пресеты для синтезаторов",
-    "Multitracks",
-    "Материалы для создания меню и обложек DVD",
-    "Стили, кисти, формы и узоры для Adobe Photoshop",
-    "Шрифты",
-    "Разное (Материалы для мультимедиа и дизайна)"
-  ],
-  "ГИС, системы навигации и карты": [
-    "ГИС (Геоинформационные системы)",
-    "Карты, снабженные программной оболочкой",
-    "Атласы и карты современные (после 1950 г.)",
-    "Атласы и карты старинные (до 1950 г.)",
-    "Карты прочие (астрономические, исторические, тематические)",
-    "Встроенная автомобильная навигация",
-    "Garmin",
-    "Ozi",
-    "TomTom",
-    "Navigon / Navitel",
-    "Igo",
-    "Разное - системы навигации и карты"
-  ],
-  "Приложения для мобильных устройств": [
-    "Приложения для Android",
-    "Приложения для Java",
-    "Приложения для Symbian",
-    "Приложения для Windows Mobile",
-    "Приложения для Windows Phone",
-    "Софт для работы с телефоном",
-    "Прошивки для телефонов",
-    "Обои и темы"
-  ],
-  "Видео для мобильных устройств": [
-    "Видео для смартфонов и КПК",
-    "Видео в формате 3GP для мобильных"
-  ],
-  "Apple Macintosh": [
-    "Mac OS (для Macintosh)",
-    "Mac OS (для РС-Хакинтош)",
-    "Программы для просмотра и обработки видео (Mac OS)",
-    "Программы для создания и обработки графики (Mac OS)",
-    "Плагины для программ компании Adobe (Mac OS)",
-    "Аудио редакторы и конвертеры (Mac OS)",
-    "Системные программы (Mac OS)",
-    "Офисные программы (Mac OS)",
-    "Программы для интернета и сетей (Mac OS)",
-    "Другие программы (Mac OS)"
-  ],
-  "iOS": [
-    "Программы для iOS",
-    "Игры для iOS",
-    "Разное для iOS"
-  ],
-  "Видео": [
-    "Фильмы для iPod, iPhone, iPad",
-    "Сериалы для iPod, iPhone, iPad",
-    "Мультфильмы для iPod, iPhone, iPad",
-    "Аниме для iPod, iPhone, iPad",
-    "Музыкальное видео для iPod, iPhone, iPad"
-  ],
-  "Видео HD": [
-    "Фильмы HD для Apple TV",
-    "Сериалы HD для Apple TV",
-    "Мультфильмы HD для Apple TV",
-    "Документальное видео HD для Apple TV",
-    "Музыкальное видео HD для Apple TV"
-  ],
-  "Аудио": [
-    "Аудиокниги (AAC, ALAC)",
-    "Музыка Lossless (ALAC)",
-    "Музыка Lossy (AAC-iTunes)",
-    "Музыка Lossy (AAC)",
-    "Музыка Lossy (AAC) (Singles, EPs)"
-  ],
-  "Разное (раздачи)": [
-    "Психоактивные аудиопрограммы",
-    "Аватары, Иконки, Смайлы",
-    "Живопись, Графика, Скульптура, Digital Art",
-    "Картинки",
-    "Любительские фотографии",
-    "Обои",
-    "Фото знаменитостей",
-    "Аудио",
-    "Музыка (lossy)",
-    "Музыка (lossless)",
-    "Видео",
-    "Публикации и учебные материалы (тексты)",
-    "Трейлеры и дополнительные материалы к фильмам",
-    "Любительские видеоклипы"
-  ],
-  "Тестовый форум": [],
-  "Отчеты о встречах": []
-}
