+++ docker-compose.yml
@@ -1,0 +1,42 @@
+version: '3.8'
+
+services:
+  postgres:
+    image: postgres:13
+    container_name: postgres_db
+    environment:
+      POSTGRES_USER: your_db_user
+      POSTGRES_PASSWORD: your_db_password
+      POSTGRES_DB: your_db_name
+    volumes:
+      - postgres_data:/var/lib/postgresql/data
+    ports:
+      - "5432:5432"
+    networks:
+      - torrent_bot_network
+
+  torrent_bot:
+    build: .
+    container_name: torrent_bot
+    environment:
+      POSTGRES_USER: your_db_user
+      POSTGRES_PASSWORD: your_db_password
+      POSTGRES_HOST: postgres
+      POSTGRES_PORT: 5432
+      POSTGRES_DB: your_db_name
+      TELEGRAM_TOKEN: your_telegram_token
+      TORRENT_FOLDER_PATH: /app/torrents
+    volumes:
+      - .:/app
+      - torrents:/app/torrents
+    depends_on:
+      - postgres
+    networks:
+      - torrent_bot_network
+
+volumes:
+  postgres_data:
+  torrents:
+
+networks:
+  torrent_bot_network:
